Fix a bug with wallet rescan progress display.

+++ src/wallet/wallet_ismine.cpp
@@ -150,9 +150,9 @@ CBlockIndex* CWallet::ScanForWalletTransactions(CBlockIndex* pindexStart, bool f
             // Temporarily release lock to allow shadow key allocation a chance to do it's thing
             LEAVE_CRITICAL_SECTION(cs_main)
             LEAVE_CRITICAL_SECTION(cs_wallet)
-            nTransactionScanProgressPercent = ((pindex->nHeight-nProgressStart) / (nWorkQuantity)) * 100;
+            nTransactionScanProgressPercent = ((pindex->nHeight-nProgressStart) / (double)(nWorkQuantity)) * 100;
             nTransactionScanProgressPercent = std::max(1, std::min(99, nTransactionScanProgressPercent));
-            if (pindex->nHeight % 100 == 0 && nProgressTip - nProgressStart > 0.0)
+            if (pindex->nHeight % 100 == 0 && nProgressTip - nProgressStart > 0)
             {
                 ShowProgress(_("Rescanning..."), nTransactionScanProgressPercent);
             }
